temp: refactor callmessage classes
