+++ src/CallInvite.tsx
@@ -20,8 +20,12 @@ import type {
   NativeCallInviteMessageReceivedEvent,
 } from './type/CallInvite';
 import type { CustomParameters, Uuid } from './type/common';
-import { CallMessage } from './CallMessage';
-import { OutgoingCallMessage } from './OutgoingCallMessage';
+import {
+  CallMessage,
+  validateCallMessage,
+} from './CallMessage/BaseCallMessage';
+import { IncomingCallMessage } from './CallMessage/IncomingCallMessage';
+import { OutgoingCallMessage } from './CallMessage/OutgoingCallMessage';
 import { Constants } from './constants';
 
 /**
@@ -62,7 +66,7 @@ export declare interface CallInvite {
   /** @internal */
   emit(
     messageReceivedEvent: CallInvite.Event.MessageReceived,
-    callMessage: CallMessage
+    incomingCallMessage: IncomingCallMessage
   ): boolean;
 
   /**
@@ -370,7 +374,7 @@ export class CallInvite extends EventEmitter {
   ) => {
     const { callMessage: callMessageInfo } = nativeCallInviteEvent;
 
-    const callMessage = new CallMessage(callMessageInfo);
+    const callMessage = new IncomingCallMessage(callMessageInfo);
 
     this.emit(CallInvite.Event.MessageReceived, callMessage);
   };
@@ -558,9 +562,7 @@ export class CallInvite extends EventEmitter {
    *    - Rejects when the message is unable to be sent.
    */
   async sendMessage(message: CallMessage): Promise<OutgoingCallMessage> {
-    const content = message.getContent();
-    const contentType = message.getContentType();
-    const messageType = message.getMessageType();
+    const { content, contentType, messageType } = validateCallMessage(message);
 
     const voiceEventSid = await NativeModule.call_sendMessage(
       this._uuid,
@@ -744,6 +746,8 @@ export namespace CallInvite {
      * @remarks
      * See {@link (CallInvite:interface).(addListener:5)}.
      */
-    export type MessageReceived = (callMessage: CallMessage) => void;
+    export type MessageReceived = (
+      incomingCallMessage: IncomingCallMessage
+    ) => void;
   }
 }
