+++ src/CallMessage/IncomingCallMessage.ts
@@ -6,41 +6,8 @@
  */
 
 import { EventEmitter } from 'eventemitter3';
-import type { NativeCallMessageInfo } from './type/CallMessage';
-import { InvalidArgumentError } from './error/InvalidArgumentError';
-
-/**
- * The constituent values of a Call Message.
- *
- * @public
- */
-export interface CallMessage {
-  /**
-   * The content of the message. This value should match the content type
-   * parameter.
-   *
-   * See {@link CallMessage.contentType} for more information.
-   */
-  content: string;
-
-  /**
-   * The content type of the message. This value should accurately describe
-   * the content of the message. The following values are accepted:
-   *
-   * - "application/json"
-   *
-   * If no value is defined, then the default value of "application/json" will
-   * be used.
-   */
-  contentType?: string;
-
-  /**
-   * The message type. The following values are accepted:
-   *
-   * - "user-defined-message"
-   */
-  messageType: string;
-}
+import type { NativeCallMessageInfo } from '../type/CallMessage';
+import { validateCallMessage } from './BaseCallMessage';
 
 /**
  * CallMessage API is in beta.
@@ -131,44 +98,4 @@ export class IncomingCallMessage extends EventEmitter {
   getSid(): string | undefined {
     return this._voiceEventSid;
   }
-}
-
-/**
- * Parse CallMessage values. Used when constructing a CallMessage from the
- * native layer, or by the Call and CallInvite classes when sending a
- * CallMessage.
- *
- * @param message the CallMessage details.
- *
- * @internal
- */
-export function validateCallMessage(message: {
-  content: any;
-  contentType?: string;
-  messageType: string;
-}) {
-  const content = message.content;
-  const messageType = message.messageType;
-
-  let contentType = message.contentType;
-
-  if (typeof contentType === 'undefined') {
-    contentType = 'application/json';
-  }
-
-  if (typeof contentType !== 'string') {
-    throw new InvalidArgumentError(
-      'If "contentType" is present, it must be of type "string".'
-    );
-  }
-
-  if (typeof messageType !== 'string') {
-    throw new InvalidArgumentError('"messageType" must be of type "string".');
-  }
-
-  if (typeof content === 'undefined' || content === null) {
-    throw new InvalidArgumentError('"content" must be defined and not "null".');
-  }
-
-  return { content, contentType, messageType };
 }
