+++ src/CallMessage/OutgoingCallMessage.ts
@@ -5,16 +5,16 @@
  * See LICENSE in the project root for license information.
  */
 
-import type { TwilioError } from './error';
+import type { TwilioError } from '../error';
 import type {
   NativeCallMessageEvent,
   NativeCallMessageEventType,
   NativeCallMessageInfo,
-} from './type/CallMessage';
-import { Constants } from './constants';
-import { NativeEventEmitter } from './common';
-import { constructTwilioError } from './error/utility';
-import { CallMessage } from './CallMessage';
+} from '../type/CallMessage';
+import { Constants } from '../constants';
+import { NativeEventEmitter } from '../common';
+import { constructTwilioError } from '../error/utility';
+import { IncomingCallMessage } from './IncomingCallMessage';
 
 /**
  * Defines strict typings for all events emitted by {@link (OutgoingCallMessage:class)
@@ -44,12 +44,6 @@ export declare interface OutgoingCallMessage {
 
   /** @internal */
   emit(sentEvent: OutgoingCallMessage.Event.Sent): boolean;
-
-  /** @internal */
-  emit(
-    outgoingCallMessageEvent: OutgoingCallMessage.Event,
-    ...args: any[]
-  ): boolean;
 
   /**
    * ----------------
@@ -152,7 +146,7 @@ export declare interface OutgoingCallMessage {
  *
  * @public
  */
-export class OutgoingCallMessage extends CallMessage {
+export class OutgoingCallMessage extends IncomingCallMessage {
   /**
    * Handlers for native OutgoingCallMessage events. Set upon construction so we can
    * dynamically bind events to handlers.
