+++ src/index.tsx
@@ -15,8 +15,9 @@ export { Voice } from './Voice';
 export { AudioDevice } from './AudioDevice';
 export { Call } from './Call';
 export { CallInvite } from './CallInvite';
-export { CallMessage } from './CallMessage';
-export { OutgoingCallMessage } from './OutgoingCallMessage';
+export { CallMessage } from './CallMessage/BaseCallMessage';
+export { IncomingCallMessage } from './CallMessage/IncomingCallMessage';
+export { OutgoingCallMessage } from './CallMessage/OutgoingCallMessage';
 export { CustomParameters } from './type/common';
 export { CallKit } from './type/CallKit';
 export { RTCStats } from './type/RTCStats';
